+++ src/test/ts/browser/EditorInitTest.ts
@@ -69,13 +69,6 @@ describe('EditorInitTest', () => {
         TinyAssertions.assertContent(ctx.editor, '<p>New Value</p>');
       });
 
-      it('Disabled prop should disable editor', async () => {
-        using ctx = await render();
-        Assertions.assertEq('Should be design mode', true, '4' === version ? !ctx.editor.readonly : ctx.editor.mode.get() === 'design');
-        await ctx.reRender({ ...defaultProps, disabled: true });
-        Assertions.assertEq('Should be readonly mode', true, '4' === version ? ctx.editor.readonly : ctx.editor.mode.get() === 'readonly');
-      });
-
       it('Using an overriden props will cause a TS error', async () => {
         // eslint-disable-next-line @typescript-eslint/no-unused-vars
         using _ = await render({