+++ CHANGELOG.md
@@ -6,6 +6,12 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## Unreleased
 
+### Changed
+- The `disabled` property now toggles the `disabled` option. #TINY-11906
+
+### Added
+- Added `readonly` property that can be used to toggle the `readonly` mode. #TINY-11906
+
 ## 6.1.0 - 2025-03-31
 
 ### Added